% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily2fraction-method.R
\name{daily2fraction}
\alias{daily2fraction}
\title{Converter from daily Date to fraction}
\usage{
daily2fraction(d)
}
\arguments{
\item{d}{array of Date object, or string in ISO \code{yyyy-mm-dd} format}
}
\value{
numeric, year in fraction convention
}
\description{
Utility to convert from daily Date (R's Date object) to fraction.
}
\examples{
  daily2fraction(as.Date("2017-01-15")) # 2017.038
  daily2fraction(as.Date("2017-02-14")) # 2017.122
  daily2fraction(as.Date("2017-07-15")) # 2017.538

}
\author{
Stephen H. Lihn
}
\keyword{utility}
