% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction2daily-method.R
\name{fraction2daily}
\alias{fraction2daily}
\title{Converter from fraction to daily Date}
\usage{
fraction2daily(fraction)
}
\arguments{
\item{fraction}{numeric, representing year in fraction convention.}
}
\value{
array of Date object
}
\description{
Utility to convert from fraction to daily Date (R's Date object).
}
\examples{
  fraction2daily(2017.038) # 2017-01-15
  fraction2daily(2017.125) # 2017-02-15

}
\author{
Stephen H. Lihn
}
\keyword{utility}
