% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jubilee-class.R
\docType{class}
\name{jubilee-class}
\alias{jubilee-class}
\title{The jubilee class}
\description{
This S4 class stores raw and derived data,
channel regression settings and results.
}
\section{Slots}{

\describe{
\item{\code{call}}{the match.call slot.}

\item{\code{lookback.channel}}{numeric, the look-back channel in years.}

\item{\code{fwd.rtn.duration}}{numeric, the forward return duration in years.}

\item{\code{reg.dtb}}{data.table, contains the regression data.}

\item{\code{dtb}}{data.table, contains the consolidated market data.}

\item{\code{rate.spread.mean}}{numeric, the mean of the yield spread,
used to calculate \code{rate.spread.norm} column.}

\item{\code{create.time}}{\code{POSIXct}, records the creation time of this object.}
}}

\keyword{class}
