% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jubilee-fred-data-method.R
\name{jubilee.fred_data}
\alias{jubilee.fred_data}
\title{Internal utility to download time series data from FRED}
\usage{
jubilee.fred_data(symbol, col_out = "Close", retry = 3)
}
\arguments{
\item{symbol}{character, the name of the time series}

\item{col_out}{character, the name of the output closing price column. Default is "Close"}

\item{retry}{numeric, number of retries on the URL. Default is 3.}
}
\value{
The \code{xts} object for the time series
}
\description{
This utility downloads time series from FRED. Many time series that this package uses
are available on FRED. Therefore, this utility is used to provide daily or monthly updates
by concatenating live data to the internal static data.
}
\examples{
\dontrun{
   jubilee.fred_data("VIXCLS") # VIX
}

}
\keyword{data}
