% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jubilee-locate-file.R
\name{jubilee.locate_file}
\alias{jubilee.locate_file}
\title{Internal utility to locate static file}
\usage{
jubilee.locate_file(local_file, stop = TRUE)
}
\arguments{
\item{local_file}{character, the file name of an internal file.}

\item{stop}{logical, whether to stop if file can't be located. Default is \code{TRUE}.}
}
\value{
The path to the file, or else, an empty string
}
\description{
This utility returns the path to internal file
}
\examples{
  jubilee.locate_file("UNRATE.csv")

}
\author{
Stephen H. Lihn
}
\keyword{data}
