% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jubilee-optimal-tb3ms-method.R
\name{jubilee.optimal_tb3ms}
\alias{jubilee.optimal_tb3ms}
\title{Calculate the optimal TB3MS}
\usage{
jubilee.optimal_tb3ms(dtb, rs, penalty = c(1, 1, 1))
}
\arguments{
\item{dtb}{data table, usually this is \code{lm.dtb} of the \code{rs} object,
with GDP log-return percent \code{(logrp.N, logrp.K)} calculated.}

\item{rs}{the list returned from \code{jubilee.macro_fit}.}

\item{penalty}{numeric, the penalty vector for the models.
Default is \code{c(1,1,1)}.}
}
\value{
The data table containing the "optimal.tb3ms" column
}
\description{
This utility calculates the optimal TB3MS using the analytic solution.
}
\author{
Stephen H. Lihn
}
\keyword{data}
