% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jubilee-predict-method.R
\name{jubilee.predict}
\alias{jubilee.predict}
\alias{jubilee.predict_real}
\title{Make prediction based on linear regression}
\usage{
jubilee.predict(object, lm, data)

jubilee.predict_real(object, lm, data)
}
\arguments{
\item{object}{object of jubilee class}

\item{lm}{the linear model}

\item{data}{data used to predict (similar to \code{newdata} of \code{stats::predict})}
}
\value{
data.table containing the prediction
}
\description{
Make prediction based on the linear regression of the forward return.
Refer to the tutorial for more detail.
}
\references{
See Section 7 of Stephen H.T. Lihn, "Jubilee Tectonic Model:
    Forecasting Long-Term Growth and Mean Reversion in the U.S. Stock Market."
    Available at \url{http://dx.doi.org/10.2139/ssrn.3156574}
}
\author{
Stephen H. Lihn
}
\keyword{model}
