% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jubilee-repo-class.R
\docType{class}
\name{jubilee.repo-class}
\alias{jubilee.repo-class}
\title{The jubilee repository class}
\description{
This S4 class stores the raw data for the jubilee package
}
\section{Slots}{

\describe{
\item{\code{call}}{The match.call slot}

\item{\code{ie}}{data.table, contains the combined data from \code{ie.raw}, \code{ws}, and \code{inflation}.}

\item{\code{yield.inversion}}{numeric, the fractions of yield curve inversion}

\item{\code{raw.ie}}{data.table, contains the data from \code{ie_data.xls} of Robert Shiller}

\item{\code{ws}}{data.table, contains the historical market return data from William Schwert}

\item{\code{inflation}}{data.table, contains the historical inflation data from Minneapolis FED}

\item{\code{comm.int}}{data.table, contains the historical commercial interest rate}

\item{\code{tb3ms}}{data.table, contains the historical 3-month Treasury bill rate}

\item{\code{gold}}{data.table, contains the historical monthly gold prices}

\item{\code{gold2}}{data.table, contains the historical annual gold prices}

\item{\code{create.time}}{\code{POSIXct}, records the creation time of this object.}
}}

\keyword{class}
