% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jubilee-std-fault-line-method.R
\name{jubilee.std_fault_line}
\alias{jubilee.std_fault_line}
\title{Standard fault line data sets}
\usage{
jubilee.std_fault_line(name)
}
\arguments{
\item{name}{character, the name of the collection. 
If "list" is supplied, the list of names will be returned.
If a numeric array is supplied, it will be converted to a matrix format.}
}
\value{
numeric, pairs of fault lines, each is c(year, delta)
}
\description{
This method defines a collection of standard fault line data sets that have been analyzed
and optimized in the research. It is intended for end users to produce
standard regressions, forecasts, and charts quickly.
}
\examples{
  jubilee.std_fault_line("r_nom_f10_5ftr_4fl")
  jubilee.std_fault_line("r_nom_f20_5ftr_2fl")
  jubilee.std_fault_line("r_nom_f20_5ftr_2fl_ramp5y")

}
\author{
Stephen H. Lihn
}
\keyword{model}
