% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jubilee-yield-inversion-method.R
\name{jubilee.yield_inversion}
\alias{jubilee.yield_inversion}
\title{List of dates for yield curve inversion}
\usage{
jubilee.yield_inversion()
}
\value{
numeric, in the unit of fraction.
}
\description{
List of dates for yield curve inversion, generally compliant to the dating
of business cycles after WWII in the U.S.. This data is also stored in the
\code{yield.inversion} slot in the \code{jubilee.repo} object.
}
\examples{
  jubilee.yield_inversion()

}
\author{
Stephen H. Lihn
}
\keyword{constructor}
