% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri-wave-class.R
\docType{class}
\name{tri.wave class}
\alias{tri.wave class}
\title{The triangular wave model class}
\description{
This S4 class defines the parameters in the triangular wave model.
}
\section{Slots}{

\describe{
\item{\code{call}}{the match.call slot.}

\item{\code{a.t}}{numeric, the look-back channel in years}

\item{\code{a0}}{numeric, the look-back channel in years}

\item{\code{s1}}{numeric, the forward return duration in years}

\item{\code{s2}}{numeric, the start fraction of in-sample training period}

\item{\code{y.mean}}{numeric, the end fraction of in-sample training period}

\item{\code{y.amp}}{numeric, the end fraction of in-sample training period}

\item{\code{y.t}}{numeric, the end fraction of in-sample training period}

\item{\code{y.p}}{numeric, the end fraction of in-sample training period}
}}

\references{
See Section 4 of Stephen H.T. Lihn, "Jubilee Tectonic Model:
    Forecasting Long-Term Growth and Mean Reversion in the U.S. Stock Market."
    Available at \url{http://dx.doi.org/10.2139/ssrn.3156574}
}
\keyword{class}
