% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri-wave-constructor.R
\name{tri.wave}
\alias{tri.wave}
\title{Constructor of \code{tri.wave} class}
\usage{
tri.wave()
}
\value{
an object of \code{tri.wave} class
}
\description{
Construct an \code{tri.wave} object to simulate the triangular wave model.
}
\examples{
w <- tri.wave()

}
\author{
Stephen H. Lihn
}
\keyword{constructor}
