% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jubilee-package.R
\docType{package}
\name{jubilee-package}
\alias{jubilee-package}
\title{jubilee: A package to forecast long-term growth of the US stock market
         and business cycles}
\description{
The jubilee package provides the core class and functions
to forecast long-term growth of the U.S. stock market.
It also contains a module for business cycles, optimal interest rate,
and recession forecasts.
A tutorial is provided to demonstrate how to use this package
and explain the relation between the mathematical notations and
the functions and data columns in this package.
}
\references{
Stephen H.T. Lihn, "Jubilee Tectonic Model:
    Forecasting Long-Term Growth and Mean Reversion in the U.S. Stock Market."
    Available at SSRN: \url{https://ssrn.com/abstract=3156574} or
    via DOI: \url{http://dx.doi.org/10.2139/ssrn.3156574}

Stephen H.T. Lihn, "Business Cycles, Optimal Interest Rate,
    and Recession Forecast From Yield Curve, Unemployment, GDP, and Payrolls."
    Available at SSRN: \url{https://ssrn.com/abstract=3422278}
}
\author{
Stephen H-T. Lihn
}
